+++ go.mod
@@ -1,6 +1,8 @@
 module github.com/rollkit/celestia-openrpc
 
-go 1.20
+go 1.21
+
+toolchain go1.21.3
 
 require (
 	cosmossdk.io/math v1.1.2
