+++ types/das/api.go
@@ -1,0 +1,10 @@
+package das
+
+import (
+	"context"
+)
+
+type API struct {
+	SamplingStats func(ctx context.Context) (SamplingStats, error) `perm:"read"`
+	WaitCatchUp   func(ctx context.Context) error                      `perm:"read"`
+}
